requirements.txt
annotated-types==0.7.0
anyio==4.3.0
certifi==2024.6.2
click==8.1.7
fastapi==0.108.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
MarkupSafe==2.1.5
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pluggy==1.5.0
pydantic==2.7.1
pydantic_core==2.18.2
pytest==8.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.32.0.post1
typing_extensions==4.11.0
tzdata==2024.1
uvicorn==0.29.0
Werkzeug==3.0.3
